chore: removed tooltip.time
