+++ apps/docs/src/stories/Tooltip/Hover.stories.tsx
@@ -32,10 +32,17 @@ const meta: Meta<TooltipHoverProps> = {
       },
       type: { name: 'boolean', required: false },
     },
+    closeTime: {
+      control: 'number',
+      description: 'Define o tempo em milissegundos para fechar o conteúdo do Tooltip.',
+      table: {
+        type: { summary: 'number' },
+      },
+      type: { name: 'number', required: false },
+    },
     delayDuration: {
       control: 'number',
-      description:
-        'Define o tempo de delay em milissegundos para exibir o conteúdo do Tooltip.',
+      description: 'Define o tempo de delay em milissegundos para exibir o conteúdo do Tooltip.',
       table: {
         type: { summary: 'number' },
         defaultValue: { summary: '300' },
@@ -53,8 +60,7 @@ const meta: Meta<TooltipHoverProps> = {
     },
     side: {
       control: { type: 'inline-radio' },
-      description:
-        'Define o local em que o componente irá aparecer, relativo ao trigger.',
+      description: 'Define o local em que o componente irá aparecer, relativo ao trigger.',
       options: ['bottom', 'left', 'right', 'top'],
       table: {
         type: {
@@ -80,8 +86,7 @@ const meta: Meta<TooltipHoverProps> = {
     text: {
       control: 'none',
       name: 'text (deprecated)',
-      description:
-        '<s>Define o valor em texto do componente.</s> (deprecated) - Use a propriedade `content`.',
+      description: '<s>Define o valor em texto do componente.</s> (deprecated) - Use a propriedade `content`.',
       table: {
         type: { summary: 'text' },
       },
