+++ package.json
@@ -24,10 +24,10 @@
   },
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.4.0",
-    "@typescript-eslint/parser": "^6.4.0",
+    "@typescript-eslint/eslint-plugin": "^6.4.1",
+    "@typescript-eslint/parser": "^6.4.1",
     "eslint": "^8.33.0",
-    "eslint-config-next": "^13.4.7",
+    "eslint-config-next": "^13.4.19",
     "eslint-config-prettier": "^8.8.0",
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-import": "^2.27.5",
@@ -39,9 +39,9 @@
     "eslint-plugin-unused-imports": "^3.0.0",
     "glob": "^7.2.3",
     "husky": "^8.0.3",
-    "lint-staged": "^13.1.0",
+    "lint-staged": "^13.1.4",
     "phosphor-react": "latest",
-    "prettier": "2.8.3",
+    "prettier": "2.8.8",
     "prettier-plugin-tailwindcss": "^0.4.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
